+++ atan2.go
@@ -26,9 +26,9 @@ package go-math32
 //	Atan2(y<0, -Inf) = -Pi
 //	Atan2(+Inf, x) = +Pi/2
 //	Atan2(-Inf, x) = -Pi/2
-func Atan2(y, x float64) float64
+func Atan2(y, x float32) float32
 
-func atan2(y, x float64) float64 {
+func atan2(y, x float32) float32 {
 	// special cases
 	switch {
 	case IsNaN(y) || IsNaN(x):
