+++ cmplx/rect.go
@@ -7,7 +7,7 @@ package cmplx
 import "go-math32"
 
 // Rect returns the complex number x with polar coordinates r, θ.
-func Rect(r, θ float64) complex128 {
+func Rect(r, θ float32) complex128 {
 	s, c := math.Sincos(θ)
 	return complex(r*c, r*s)
 }
